chore: make the SettingCard directory and move relative files
