+++ collision1/src/App.tsx
@@ -1,6 +1,6 @@
 import { Button, DatePicker, Space, version } from "antd";
 import { useState } from "react";
-import SettingCard from "./components/SettingCard";
+import SettingCard from "./components/SettingCard/SettingCard";
 
 function App() {
   return (
