+++ collision1/src/components/SettingCard/SettingInput.tsx
